Add github icon with link
